requirements.txt
beautifulsoup4==4.14.3
chromadb==1.3.5
dotenv==0.9.9
httptools==0.7.1
langchain==1.1.2
langchain-community==0.4.1
langchain-openai==1.1.0
pypdf==6.4.0
sentence-transformers==5.1.2
watchfiles==1.1.1
websockets==15.0.1
pandas==2.2.3
python-pptx==0.6.23
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.9.2
python-multipart==0.0.9
